move builtin_type_shadow to its own module
