+++ tests/ui/builtin_type_shadow.rs
